"devLocationsOfInterest" API oute naming issue fixed
Previously, there were two branches of API routes: "LocationsOfInterest" and "dev_LocationsOfInterest".
One was a development database and one was production. This should be done with environment variables rather than different URLs. This commit has removed all of "devLocationsOfInterest" for consistency.
